// packages/fhir-r5/src/v5.0.0/examplescenarioactor/types.ts
import type { Extension, Element } from "../core/types";

/** Generated from FHIR JSON Schema */

/** Example of workflow instance. */
export interface ExampleScenarioActor {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  key?: string;

  _key?: Element;

  type?: string;

  _type?: Element;

  title?: string;

  _title?: Element;

  description?: string;

  _description?: Element;
}